test(models): satisfy coverage test layout gates
